+++ packages/block-editor/src/components/provider/use-media-upload-settings.js
@@ -23,6 +23,7 @@ function useMediaUploadSettings( settings = {} ) {
 			bigImageSizeThreshold: settings.bigImageSizeThreshold,
 			imageStripMeta: settings.imageStripMeta,
 			imageMaxBitDepth: settings.imageMaxBitDepth,
+			animatedImageSubsizes: settings.animatedImageSubsizes,
 		} ),
 		[ settings ]
 	);