+++ phpunit/media/media-processing-test.php
@@ -123,6 +123,7 @@ class Media_Processing_Test extends WP_UnitTestCase {
 		$this->assertArrayNotHasKey( 'image_sizes', $data );
 		$this->assertArrayNotHasKey( 'image_strip_meta', $data );
 		$this->assertArrayNotHasKey( 'image_max_bit_depth', $data );
+		$this->assertArrayNotHasKey( 'animated_image_subsizes', $data );
 	}
 
 	/**
@@ -158,6 +159,8 @@ class Media_Processing_Test extends WP_UnitTestCase {
 		$this->assertTrue( $data['image_strip_meta'] );
 		$this->assertArrayHasKey( 'image_max_bit_depth', $data );
 		$this->assertSame( 16, $data['image_max_bit_depth'] );
+		$this->assertArrayHasKey( 'animated_image_subsizes', $data );
+		$this->assertFalse( $data['animated_image_subsizes'] );
 	}
 
 	/**
@@ -197,6 +200,23 @@ class Media_Processing_Test extends WP_UnitTestCase {
 		$data    = $index->get_data();
 
 		$this->assertSame( 8, $data['image_max_bit_depth'] );
+	}
+
+	/**
+	 * @covers ::gutenberg_media_processing_filter_rest_index
+	 */
+	public function test_get_rest_index_honors_animated_image_subsizes_filter() {
+		wp_set_current_user( self::$admin_id );
+
+		add_filter( 'wp_generate_animated_image_subsizes', '__return_true' );
+
+		$server = new WP_REST_Server();
+
+		$request = new WP_REST_Request( 'GET', '/' );
+		$index   = $server->dispatch( $request );
+		$data    = $index->get_data();
+
+		$this->assertTrue( $data['animated_image_subsizes'] );
 	}
 
 	/**