+++ packages/components/src/private-apis.ts
@@ -24,15 +24,18 @@ import {
 	ValidatedNumberControl,
 	ValidatedSelectControl,
 	ValidatedRadioControl,
+	ValidatedContentEditableControl,
 	ValidatedTextControl,
 	ValidatedTextareaControl,
 	ValidatedToggleControl,
 	ValidatedToggleGroupControl,
 } from './validated-form-controls';
 import { ValidatedFormTokenField } from './validated-form-controls/components/form-token-field';
+import ContentEditableControl from './content-editable-control';
 
 export const privateApis = {};
 lock( privateApis, {
+	ContentEditableControl,
 	__experimentalPopoverLegacyPositionToPlacement,
 	ComponentsContext,
 	Tabs,
@@ -52,6 +55,7 @@ lock( privateApis, {
 	ValidatedNumberControl,
 	ValidatedSelectControl,
 	ValidatedRadioControl,
+	ValidatedContentEditableControl,
 	ValidatedTextControl,
 	ValidatedTextareaControl,
 	ValidatedToggleControl,