+++ packages/eslint-plugin/rules/components-no-missing-40px-size-prop.js
@@ -15,9 +15,6 @@ const COMPONENTS_REQUIRING_40PX = new Set( [
 	'Button',
 	'ClipboardButton',
 	'IconButton',
-	'InputControl',
-	'NumberControl',
-	'SelectControl',
 ] );
 
 module.exports = {